requirements-test.txt
pytest>=7.4.0
pyyaml>=6.0
